get rid of a few more warnings
git-svn-id: svn+ssh://gate.spectrum.uni-bielefeld.de/homes/tbaumann/inpro-svnrepo/trunk@2810 19460358-522e-0410-942e-fdb976ae3ca2

+++ src/gov/nist/sphere/jaudio/SphereFileReader.java
@@ -99,7 +99,7 @@ public class SphereFileReader extends AudioFileReader {
         int channelCount = header.getChannelCount();
         int resolution = header.getSampleNBytes() * 8;
         boolean isBigEndian = header.isDataBigEndian();
-        format = new AudioFormat(encoding, (float) sampleRate, resolution,
+        format = new AudioFormat(encoding, sampleRate, resolution,
                 channelCount, channelCount * header.getSampleNBytes(), sampleRate, isBigEndian);
       } catch (SphereException e) {
     	e.printStackTrace();
